modernize race conditions test
